refactor(server): move the nine ai_gateway modules out of server/agent_auth (git mv, move-only)
