+++ server/proliferate/server/agent_auth/__init__.py
@@ -1,1 +1,1 @@
-"""Agent LLM gateway enrollment domain (LiteLLM)."""
+"""Agent credential selections, key vault, state delivery, and org agent policy."""
