+++ server/proliferate/server/ai_gateway/__init__.py
@@ -1,0 +1,1 @@
+"""Agent LLM gateway enrollment domain (LiteLLM)."""
