+++ server/tests/integration/test_agent_gateway_enrollment.py
@@ -39,9 +39,9 @@ from proliferate.db.store.billing_subjects import (
     ensure_personal_billing_subject,
 )
 from proliferate.integrations.litellm import LiteLLMIntegrationError, LiteLLMVirtualKey
-from proliferate.server.agent_auth import enrollment as enrollment_service
-from proliferate.server.agent_auth.budget import get_gateway_enrollment_for_user
-from proliferate.server.agent_auth.enrollment import (
+from proliferate.server.ai_gateway import enrollment as enrollment_service
+from proliferate.server.ai_gateway.budget import get_gateway_enrollment_for_user
+from proliferate.server.ai_gateway.enrollment import (
     _parse_budget,
     _remaining_credit_budget_raw,
     backfill_enrollments,
