+++ server/tests/integration/test_agent_gateway_topup_fixes.py
@@ -19,9 +19,9 @@ from proliferate.db.store.billing_subjects import (
     ensure_organization_billing_subject,
     set_billing_subject_overage_enabled,
 )
-from proliferate.server.agent_auth import usage_import as usage_import_service
-from proliferate.server.agent_auth.enrollment import ensure_org_enrollment
-from proliferate.server.agent_auth.topups import (
+from proliferate.server.ai_gateway import usage_import as usage_import_service
+from proliferate.server.ai_gateway.enrollment import ensure_org_enrollment
+from proliferate.server.ai_gateway.topups import (
     create_llm_topup_grant,
     run_llm_topups,
 )
