+++ server/tests/unit/test_background_workers_flag.py
@@ -12,7 +12,7 @@ from __future__ import annotations
 import pytest
 
 from proliferate.config import settings
-from proliferate.server.agent_auth.worker import (
+from proliferate.server.ai_gateway.worker import (
     start_agent_gateway_enrollment_backfill,
     start_agent_gateway_llm_topups,
     start_agent_gateway_usage_import,
